Move static functions to inline headerfiles
