+++ Makefile
@@ -1,3 +1,3 @@
 format:
-	Tools/fix_style.sh geometric_controller/src
-	Tools/fix_style.sh trajectory_publisher/src
+	Tools/fix_code_style.sh geometric_controller/src
+	Tools/fix_code_style.sh trajectory_publisher/src
